food_recognition/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
numpy==2.1.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tensorflow